\name{closeStrategy}
\alias{closeStrategy}
\alias{closeStrategy}
\title{
    closeStrategy
}
\description{
    This connects to Machina web service, and closes the open strategy.
}
\usage{
    closeStrategy(verbose = FALSE)
}
\arguments{
    \item{verbose}{
        Not required, whether to produce verbose output (default = FALSE)
	}
}
\details{
    This connects to Machina web service, and closes the open strategy.
}
\value{
    This function closes the open strategy.
}
\author{
    Michael Gursky
}
\examples{
   # closeStrategy()
}
\keyword{Machina strategy}