% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{get_ast}
\alias{get_ast}
\title{Create an AST from R code
See: https://stackoverflow.com/a/60083792}
\usage{
get_ast(expr)
}
\arguments{
\item{expr}{an R expression}
}
\description{
Create an AST from R code
See: https://stackoverflow.com/a/60083792
}
