\name{getStrategy}
\alias{getStrategy}
\alias{getStrategy}
\title{
    getStrategy
}
\description{
    This connects to the Machina web service, and gets all the rows in the open strategy.
}
\usage{
    getStrategy(verbose = FALSE)
}
\arguments{
    \item{verbose}{
        Not required, whether to produce verbose output (default = FALSE)
	}
}
\details{
    This connects to Machina web service, and gets all the rows in the open strategy.
}
\value{
    This function and gets all the rows in the open strategy.
}
\author{
    Michael Gursky
}
\examples{
  #  getStrategy()
}
\keyword{Machina get strategy}