\name{listStrategies}
\alias{listStrategies}
\alias{LISTSTRATEGIES}
\title{
    listModels
}
\description{
    This connects to the Machina web service, and gets a list of your models.
}
\usage{
    listStrategies(verbose = FALSE)
}
\arguments{
    \item{verbose}{
        Not required, whether to produce verbose output (default = FALSE)
	}
}
\details{
    This connects to Machina web service, and and gets a list of your strategies.
}
\value{
    This function returns a list of your models.
}
\author{
    Michael Gursky
}
\examples{
  #  listStrategies()
}
\keyword{Machina list strategies}