\name{viewStrategy}
\alias{viewStrategy}
\alias{viewStrategy}
\title{
    viewStrategy
}
\description{
    This outputs the open strategy rows in a data frame.
}
\usage{
    viewStrategy(verbose = FALSE)
}
\arguments{
    \item{verbose}{
        Not required, whether to produce verbose output (default = FALSE)
	}
}
\details{
    This outputs the open strategy rows in a data frame.
}
\value{
    This outputs the open strategy rows in a data frame.
}
\author{
    Michael Gursky
}
\examples{
   # viewStrategy()
}
\keyword{Machina view strategy}